Bump allowed number of queries
During DestroyAllExpiredService.

The new query which is causing the number of queries to cross the limit
is:

`UPDATE "ci_job_artifacts" SET "expire_at" = NULL WHERE 1=0`

The where clause looks funny because there are zero wrongly expired
artifacts, yet `update_all` is called anyway, on a relation built with
an empty array. The new query is called at
app/services/ci/job_artifacts/destroy_batch_service.rb:142.

We are modifying only the test in this commit, since it is low risk.