Revert "Use OS tempdir for artifact metadata"
This reverts commit a5b0a030c38acd6da5f66f094a40882f7d2eb892.

The Rails multipart middleware only allows temporary uploads to be
stored in certain paths. This middleware includes Ruby's `Dir.tmpdir`,
but this method may not necessarily return `/tmp`
(https://github.com/ruby/ruby/blob/v2_7_5/lib/tmpdir.rb#L23-L30). For
example, if the permissions of `/tmp` do not contain the sticky bit,
Ruby may return the current working directory.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/87255 changed
the behavior so that artifact metadata is always stored in
`/tmp`. When Ruby's `Dir.tmpdir` differs from that path, then artifact
uploads will be rejected with a 400 Bad Request.

To avoid this issue, we revert the changes in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/87255 and always
take the recommended temp path from Rails.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/90083 contains
an adapted revert for `master`, while this commit is a straight revert
that targets the `15-0-stable-ee` branch.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/363701