Fix trigger to release-environments deployments
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2696

In MR https://gitlab.com/gitlab-org/gitlab/-/merge_requests/117212
we added a trigger from stable branch pipelines to do a downstream
trigger pipeline in the release-environment. However, the URL to
the repo to trigger was incorrect leading to a project not found
error. This change corrects this.