Sort admin runners page by runner version
Add the backend capabilities for the runner sorting page to
accept a new GraphQL parameter: version_prefix.

This will accept a string like `15.0.` or `14.11.126` and look
to match the runner-manager version that is being kept in the database.

This field should be marked as an `alpha` field as it is
experimental.

Add each of the version search indexes suggested during review as a
migration to the database. Update the testing to account for each
type of search.

To take advantage of the added indexes, use substrings in the
with_version_prefix function for querying the database. Automatically
add the `.` when the matched regular expression calls for it.

Changelog: added

Co-authored-by: Joseph Snyder <joe.snyder@kitware.com>