Fix no_proxy not working when DNS rebinding protection enabled
When a URL request is made, the `Gitlab::UrlBlocker` validates whether
the request should be allowed to proceed and returns the URI and
hostname that should be used to connect to the URL.

In GitLab 15.11,
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/115343
re-introduced the skipping DNS rebinding checks if a proxy server is
enabled. It also carved out an exception for the no_proxy setting: if
the URL is configured to bypass a proxy, `UrlBlocker` would allow the
DNS rebinding protection to go ahead. DNS rebinding protection
resolves the hostname to the IP address before the request is
initiated.

When the HTTP request were initiated, the `Net::HTTP` request would
compare the resolved IP against the `no_proxy` list. If the IP address
did not match any entries in `no_proxy`, the request would be
erroneously be routed to the proxy server.

This change fixes the issue of `no_proxy` being ignored by returning a
boolean, `us_proxy`, in `UrlBlocker`. If `false`, this tells the
caller whether the proxy should be disabled, since the `UrlBlocker`
normally does the necessary checks to determine whether a proxy should
be used. This commit also refactors the return value to use a Struct.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/410636

Changelog: fixed