Add task to fix migrations for 15.11 upgrades
This is specific to 15.11.x - a migration squash that is present in
15.11.0-15.11.2 (inclusive) removed a crucial migration that was
   chronologically out of order from the squash point which introduced
an error. Then the migration squash was reverted for 15.11.3, which
means that users upgrading from an earlier 15.11 would see new
migrations that haven't run yet, and an attempt to run them would error
out.

This rake task detects that condition and patches in the necessary
timestamps.