Fix crash when LDAP CA file set outside tls_options
Previously if `ldap_servers.ca_file` were set instead of
`ldap_servers.tls_options.ca_file`, the changes in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/122888 would
flag the call to `Gitlab::Options#delete`.

Unfortunately, the introduction of
`Gitlab::ErrorTracking.track_and_raise_for_dev_exception` in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/124396 caused
the process to crash with an undefined `devise` method in the `User`
module. This happened because the method attempted to load the `User`
for the application context, but since Devise is initialized in
`8_devise.rb`, this came too late for the LDAP config parsing in
`1_settings.rb`.

To avoid this error, avoid loading the `ApplicationContext` by using
`Gitlab::AppJsonLogger`.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/419485

Changelog: fixed