Prevent pipeline creation while import is running
Because the imports don't get the internal ids from the
database sequence, creating new pipelines while the
import is still running will attempt to insert
duplicate iid values in the ci_pipelines table.
This fix prevents new pipelines from being created
while the import is running. After the import is
completed, the internal_ids row for the project will
be deleted and the new pipeline will recreate it
with the appropriate value.

Changelog: fixed