Add merged_commit_sha to queries for merge requests
This column was recently added, but not to any queries because the index
did not exist yet. This index has now been added, so we can use it in
queries. This means that non-squashed fast-forward merge train merges
can now be traced back to the source MR.

See https://gitlab.com/gitlab-org/gitlab/-/issues/425100

Changelog: fixed