Remove hasEnvScopeQuery prop from ci_variable_list
`hasEnvScopeQuery` is always true for groups and projects,
so the usage of this flag is redundant now, except in the
`CiVariableShared` component where we want to skip the
environments query if none is provided. This only applies
to instance variables, since CI variables on the instance-level
are not associated with environments.

In the past, `hasEnvScopeQuery` was used to determine if we need
to do the search for environments scopes via GraphQL or on the
client-side (where we manually filter from the list of environment
scopes we initially fetched from various sources).

Since group and project CI variables use GraphQL now, the search
will always be done using the GraphQL query. We no longer need to
do a client-side search anymore.