Remove `maxlength` limit from the group description textarea
In the `General Settings` page one can add an optional group
description. Previously, this textarea field was set to a maximum of
250 characters. In this commit we have remove that limit.

There is an actual limit of 2000 imposed by the backend. If one tries to
save a group description longer than 2000 characters a warning will
appear after clicking `Save changes`.

Changelog: changed