Add LATERAL query to Security::FindingsFinder
This change adds a new query strategy using LATERAL to the
`Security::FindingsFinder` .

Currently the performance on this finder is very bad on pipelines with
large numbers of findings as it must filter all of the findings before
ordering the result.

This change allows a reduction on the number of rows filtered when
paginating the lower pages of the result.

The new query has test parity with the old one and is behind a new
feature flag - security_findings_finder_lateral_join