Add a QA test for malformed URLs
Because of https://github.com/puma/puma/pull/3094, Puma v6.4.0 returns
the status code sent from the low-level handler instead of Puma's
recommended status code. This caused malformed URLs to return a 500
error when previously it would return a 400 Bad Request error, which
skewed error metrics.

To catch this issues in the future, add a gitlab-qa test that
attempts to issue a malformed URL.

Relates to
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/16417