Do not retry a failed review-deploy job
The retries add a lot of time to the pipeline duration,
they count in the overall success rate of
review-apps, and they don't solve the initial issue
around 40% of the time.

I believe this is due to the fact that the retried job
directly tries to delete/recreate the namespace,
and the operation might still be in progress.
Having it retried manually would help have a
successful deployment, as there would be an
additional delay for it.