Update ChatName activity timestamp throttle
`ChatName` records do not have their activity timestamp correctly
updated. This is due to an incorrect time comparison. This MR replaces
`>` with `.before?`, which should yield the correct result, and updates
the specs accordingly.

In addition, it removes the global lock on the chat name, as this
behaviour duplicates the timezone check.

Changelog: fixed