Fix flaky epic spec
This spec fails at the start of a new calendar quarter because of the
way the epics are initialized with a relative due date in the past. The
epic roadmap renders the current calendar quarter only, so these epics
don't appear in the first 10-20 days of a new calendar quarter and the
expectation to find epics instead of an empty state fails.

This change sets the due date in the future for the epic fixtures so
that they will always appear in the roadmap for the current quarter.