Improve predictive jobs parallelization
Instead of using the average test file duration for a given test level
type, we now use the duration of the actual test files that will run in
the predictive jobs. That way, if we only run very slow test files, the
computation will take it into account and parallelize on more jobs.

Signed-off-by: Rémy Coutable <remy@rymai.me>