Displays ai_features token only when available
ai_features token scope was introduced to allow usage of Ai features
with a less powerful scope than API. But it is not always applicable,
since ai features need to be enabled in the instance and need to be part
of the license. Creation of ai_features token is possible for users that
have at least one group with ai features enabled, and for projects and
groups that have ai_features enabled.

Changelog: added
EE: true