Truncate diffs further to fit input token limit
Based on testing, truncating diff to 2000 words will lead to more
than 8k input tokens. As a result, we're getting random characters,
letters, or numbers as response when asking Vertex to summarize
merge request.

To fix, we reduce the words to 750. Based on testing, this is
within the limit of 8,192 input tokens.