Created new merge checks component
This merge checks component renders a collapsed list of all
the different merge checks that gets returned from the backend.
Each of these items in the merge checks will either render the `message.vue`
component which simply just renders some text.
Or it can be overriden in the `merge_checks.vue` component
by adding an item to the `COMPONENTS` object with the key
name being the same identifier from the backend.

https://gitlab.com/gitlab-org/gitlab/-/issues/423111