Fix target_branch_path in merge request sticky header
The target branch for a merge request can be displayed in one
of two places:

  * in the regular header, which is shown when the page is
    initially loaded;

  * in the more compact sticky header, shown after scrolling
    down the page a bit.

In the sticky header version, however, the target branch is
assumed to be part of the source repository instead of the target
repository. As a result, the link to the target branch is
initially correct, but once you scroll down the page and the
sticky header takes over it starts pointing to the wrong branch.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>