Fix vulnerability statistics calculation query
As we have introduced a change to create vulnerabilities with
`present_on_default_branch` as FALSE, the previous query was counting
the vulnerabilities that do not exist on the default branch as well.

With this change, we will start querying the `vulnerability_reads`
table which contains only the vulnerabilities that exist on the default
branch.

Changelog: fixed
EE: true