Use `InOperatorOptimization` for group vulnerabilities
Background Context
---------------------------

We have an issue where the report export for all vulnerabilities is
timing out.

The existing export flow is:

1. A `Vulnerability::Export` record is created.
2. A `VulnerabilityExports::ExportWorker` job is started to generate
   the exported file.
3. The exporter runs using the query generated by
   `Security::VulnerabilitiesFinder(group).execute.with_findings_scanner_identifiers_and_notes`
   and uses `CsvBuilder` to make the CSV.

The query should run in batches of 1000 via `find_each`, so we were
surprised it was creating PG timeouts

exporting the `gitlab-org` group (~40,000 vulnerabilities) seems to
cause the timeout. The original reporter of the bug has around ~320k
vulnerabilities

Proposed Fix
---------------------------

There is a nested `IN` query to find all of the `Project` IDs under a
give `Group` (and subgroups). After we have the IDs, the query should
be fairly straightforward as we have a `project_id` column on the
`vulnerabilities` table.

We want to try using the [`InOperatorOptimization` module][0] to
improve that portion of the query.

In this commit we add a code-path using that module, gated by a
FF (disabled by default).

We also need to add an index before attempting to use the module. In
this commit we schedule an async index addition, and will need to
follow-up with a synchronous index migration before toggling the FF

[0]:https://docs.gitlab.com/ee/development/database/efficient_in_operator_queries.html#the-inoperatoroptimization-module

Changelog: changed
EE: true