Update password reset task and docs
Users that are provisioned via OmniAuth have their passwords
automatically set and the `password_automatically_set` attribute
is set to true to prevent sign-in via that random password.
If an admin resets a user's password via the Rails console or
Rake task and this attribute isn't set to false, the user will
still be unable to sign in with that password. This change updates
these two processes.