Extract GitHub import helpers to module
This is a small refactor to extract the Github import API helpers to
their own module.

The motivation is to make the module more easily testable in isolation,
as a related piece of work I'm doing (the removal of the feature flag)
mentioned in this MR resulted in an `rspec:undercoverage` failure on
the #client method.