Fix dummy email for omniauth integrations
When Microsoft Entra ID (Formally azure AD) is integrated with GitLab
with OIDC flow,  `email` attribute may not contain the value even if
email scope is given.

Also the username field can have email like field
`GitLab@gitlabsandbox.onmicrosoft.com`. So when GitLab attempts
to create a temporary email for this user, it will be
`temp-email-for-oauth-gitlab@gitlabsandbox.onmicrosoft.com@gitlab.localhost.

Since this email contain two @, system will mark this user as invalid
and prevents the sign in.

Proposed solution is to remove the domain field from the user name
if it is of the form something@domain. Old behaviour is kept intact
for LDAP login as there is an existing spec which expects username
in something@domain format

Changelog: fixed