Adds revert template for security MRs
On the previous security release, two security fixes had to be reverted
because they impacted GitLab.com availability or introduced bugs to the
codebase. Links:

* https://gitlab.com/gitlab-org/gitlab/-/issues/393774#note_1331187212
* https://gitlab.com/gitlab-com/gl-infra/production/-/issues/8606#note_1335427849

The apprach for each of them dependend on some variables. This MR
template aims to describe the different options based on those
variables, and to serve as guidance for stage teams when they need to
revert a security merge request.