Configure retries for read-only Gitaly RPCs
Currently our Ruby Gitaly gRPC client will automatically perform
'transparent' retries[0], where the request has reached gRPC's internal
loadbalancer but not gone onto the wire. However, gRPC can retry
requests in a larger number of scenarios when configured to do so.

Add a `retryPolicy` to Gitaly's service_config to allow retries for any
read-only RPC that fails with an `UNAVAILABLE` status code. This status
code used exclusively to indicate a connection/network failure. We
allow up to two additional requests to be sent, at 250ms and 500ms
intervals. The overall request deadline is still honored as well. This
allows the client to handle momentary service interruptions without
bubbling errors up to users.

Gitaly sets a `MethodOption` named `op_type` on RPCs to indicate which
ones will modify the repository. This is accessible to Golang clients,
but with Ruby we are unfortunately forced to manually list RPCs known to
be read-only as the Ruby protobuf implementation does not support
accessing `MethodOption`s.
https://github.com/protocolbuffers/protobuf/issues/1198 is open to track
adding this feature to gRPC-Core.

[0] https://github.com/grpc/proposal/blob/master/A6-client-retries.md#transparent-retries