Allow admin to create identity verification exemption
When an identity verification exemption is created, a user just needs
to confirm an email address to confirm their identity. They do not need
to confirm their identity with a phone number and/or a credit card.

This also removes the phone number exemption button from the
admin interface as that has become redundant.

Changelog: added
EE: true