Implement mergeabilityChecks GraphQL field
Frontend will need to display failed merge checks on the merge
widget. In order for frontend to be able to do that, they need to
be able to query mergeability checks status via GraphQL and determine
which ones failed or not.

This adds a new `mergeabilityChecks` field for the `MergeRequest` type
which can be queried by something like:

```
mergeRequest(iid: 1) {
  mergeabilityChecks {
    identifier
    status
  }
}
```

`identifier` is the name of the mergeability check and `status`
can be `success`, `failed` or `inactive`.

`inactive` status means that the merge check didn't get executed
for the merge request due to a setting being disabled.

Changelog: added