Fix gitlab_username_claim support for OmniAuth providers
Prior to this fix gitlab_username_claim was not always working
correctly. When loading the provider config, the auth hash class
used an ivar which may not yet be initialized and would cause
false negative config lookup. This fix replaces ivar usage with
the method that will correctly get the provider name and allow
config to be found.

Changelog: fixed