Fix pipeline schedules view when owner is nil
The pipeline schedule Vue refactor in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/128913 caused
pipeline schedules not to be rendered when an owner of a pipeline
schedule was nil. This can happen if the user is deleted.

Update the GraphQL schema and Vue code to allow for a nil owner.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/427747

Changelog: fixed