Refactor Todos::Destroy services to be more concrete
We move away from the reliance on base class methods, as it makes it
hard to refactor. It was hard to refactor because part of the
application logic was in the base class, part of the logic was in the
concrete class.

After refactoring, it is clear that the `authorized_users` method never
applied to all descendant classes - it only applied to one descendant.

Similarly the `without_authorized` method only applied to a few
descendant claseses, but not all.