Ensure no seat workers run after FF disabled
Currently the seat refresh worker will run indefinitely if the FF is
modified to FALSE mid-run. This is because the worker will always think
it has more to do as the DB column that marks work as complete is no
longer updated.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/427034
EE: true