Update work item notes to use fullPath props
Work items from a project can open another work item from
a different project in a modal. Right now, fullPath is
injected which essentially comes from the url route, so
this assumes that the modal is in the same project.
We need to update the work item components to receive
fullPath by prop so that the modal can receive a fullPath
value that represents a different project.