Remove the server_side_frecent_namespaces feature flag
This removes the `server_side_frecent_namespaces` feature flag.

Going forward, GitLab instances will start tracking users' visits to
projects and groups. There is no user-facing change at the moment as
this only enables data-collection. In the next iterations, we'll start
leveraging the data to suggest "frecently" (a combination of frequency
and recency) visited namespaces to the users.

This new approach will replace the current localStorage-based one
which is flawed due to being tied to the current browser and not
calculating frecency scores smartly-enough.

Changelog: added