Fix bug with batched gitaly ref deletion duplicates
If there are duplicates sent in a single Git command we get an error.
This de-duplication works around this. We could also consider
de-duplicating in the queue but this is tricky to do without race
conditions and would also require a migration. Without the `uniq` fix
this new test fails with:

```
Gitlab::Git::CommandError:
13:unable to prepare: reference has been updated multiple times
within a transaction.
```