Do not render chunks once full message is received
We do not guarantee the order when subscription messages are received.
This is the case for chunks, but also for the complete message.

So it could happen that we receive the full message from `aiCompletion`
but then still receive one more chunks. In this case we should stop
trying to compute the full message based on the received chunks and
ignore the chunk. Otherwise this would lead to a bug where the user only
sees the last chunk rendered.