Improve `script` with `trigger` error message
The `script` and `trigger` CI keywords cannot be used together. This
commit adds a mutually exclusive validator to Entry::Processable that
ensures that the user gets an understandable error.

It also rewords the error from MutuallyExclusiveValidator for clarity.
The validator was not being used anywhere, so this is a safe change to
make

Changelog: fixed