Fix performance of the pipelines JSON endpoint
The origin of this performance problem is the pipelines tab of
merge requests. When loading the pipelines, we send a request to
"merge_requests/:id/pipelines" and return a JSON payload.
This JSON payload contains all "failed_builds". However, we don't need
this calculated array because we only need the number of failed builds.

Benchmark shows that this change improves the performance by 59%.

This change is behind
the feature flag "ci_fix_performance_pipelines_json_endpoint".