Add partition number to security findings query
All records for a particular pipeline are guaranteed to be on the same
partition in the `security_findings` table so we can add the partition
number as a parameter when querying the table for a pipeline.

This change adds:

```
... FROM security_findings WHERE security_findings.partition_number = n
```

to the query in `Security::FindingsFinder`.

As a result of this the planner only needs to visit one partition.
Without this clause all active partitions are scanner - 12 at the
time of writing.