Fix vulnerability count mismatch
Currently there is a mismatch between the vulnerability
counts on the security dashboard and vulnerability report.

The root cause is that the dashboard is using historical, trending
data, which gets populated by a cronjob.

The vulnerability report uses life, up-to-date data.
This fix uses the life data for the most current date
on the security report.

Changelog: fixed
EE: true