Add name and description denormalized columns in catalog_resources table
This MR adds the name and description columns to the catalog_resources
table. They are denormalized from the projects table. A post deploy
migration backfills the columns. The Ci::Catalog::Resource and
Project models are updated so that when a new resource is created
or when the project name/description is updated, the
corresponding columns in the catalog resources stay in sync.

Changelog: added