Set deploy key as actor for Git proxied requests (Geo)
This commit sets deploy keys as access_actor in the EE context.
Why? When a repository is not synchronized in a Geo secondary site
due to selective synchronization, and it is forced to proxy
Git commands (i.e., clone) to the primary site, the proxied request
can bypass the scope of deploy keys and perform the action
whereas deploy keys are not enabled for the project.

Changelog: fixed
EE: true