Fix two user profile related specs with new sidebar
The new sidebar adds more dropdowns to the page and therefore some CSS
selectors had to be made more specific on which dropdown to select.

Furthermore the new sidebar doesn't render a `0` pill if a user doesn't
have a follower or is following noone. The tests have been adjusted
accordingly.