Implement support for inactive mergeability check
We need to return mergeability checks as `inactive` when they're
disabled via settings/license. To do that, we need to set the
`status` of a `CheckResult` as `inactive`, when a mergeability
check service gets executed.

This adds `CheckBaseService#inactive` that will be called in
corresponding mergeability check service classes when they're
inactive.

Changelog: added