GithubImport::ObjectCounter use same TTL when read
This fixes a bug where reading the summary of counts would set a much
lower TTL for the cache key.

In large imports that take place over multiple days, this can mean
losing certain counts if someone views the summary while the import is
still taking place, as they will expire a day later, when the import is
still in progress.

https://gitlab.com/gitlab-org/gitlab/-/issues/427097#note_1615773392

Changelog: fixed