Update release and version in Sentry frontend
According to Sentry's documentation a release is a "version of your code
deployed to an environment". As gitlab.com is the source of errors for
our sentry, we should use "gon.revision" as releases.

This is importnat if we wish to integrate js sourcemaps to our workflow,
as sourcemaps artifacts will be assigned a release name that must match
the git revision in Sentry.