Add sha256 hash to NuGet symbols
When debuggers such as Visual Studio needs to download the nuget package debugging symbol files `.pdb`, they send the sha256 hash of the requested file. Thus, we need to calculate this hash an store it in the DB. When we receive the SymbolChecksum header including the sha256 hash, we match it with the one we have, and if it matches, we return the file knowing it's the correct one.

Changelog: added