Restore repositories from backup manifest files
When no backup ID is provided to gitaly-backup the latest backup is
used. For non-server-side backups this means the only backup in the tar
file.  The problem with the latest backup is that using latest is not
compatible with manifest files. Since manifest files allow us to
properly set the default branch and also to eventually support sha256
repos, here we begin sending the backup ID to gitaly-backup
unconditionally.

Now that gitaly-backup sets both the default branch and the hash format
it is not possible to tentatively create repositories that did not exist
in the backup. There is not enough information to do this. So the tests
were adjusted to suit.

Changelog: changed