Update catalog resource search query plan to use denormalized columns
The catalog_resources table now has 'name' and 'description'
columns which were denormalized from the projects table.
This MR adds a search scope to the Ci::Catalog::Resource model
and updates the query plan in Ci::Catalog::Listing
so that it no longer relies on the ProjectsFinder search
functionality. This improves the search query performance.