ADD worker and dependent services for ActivityPub subscription
This is the second part of the merging the overarching subscription MR
at https://gitlab.com/gitlab-org/gitlab/-/merge_requests/132460 .

This provides the worker performing the async resolution of inbox URLs,
and sending the Accept activity to the subscriber's inbox.

The work will be called from the inbox endpoint of our actor for
releases, when the endpoint receive a Follow activity (this part will be
added in a following MR). When it happens, the controller queues the
worker, which will be responsible of:

- Retrieving the actor's inbox URL if we're only provided the actor
  profile URL
- Retrieving the third party server shared inbox if it's provided
- Sending an Accept activity to the actor's inbox to let them know we
  received their subscription
- Updating our subscription record to mark it as active

We won't use the shared inbox yet, but we will need it as soon as we
send events when a new release is created, and the request we make to
the subscriber's profile page is where that information is made
available, so we save it right now.

The job is allowed to fail and is idempotent: restarting it if the
inboxes were already resolved won't trigger an other third party
request. If a subscription request somehow gets in for an already
existing accepted subscription, we ignore it. And if the project is not
public anymore, we delete the subscription.

The `ThirdPartyError` error has been added mainly to be nice to support,
so that they don't queue errors in the issue tracker and start analyzing
it only to realize again and again that it was just an unstable third
party server and we can't do anything about it. Hopefully the name
convey well enough the idea of "don't bother, it's not us".

After this is merged, we'll still need the controller, route and the
service creating the subscription record.