Transition to structured errors for RawBlame (step 2)
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/425441

**Problem**

A wrapper for Gitaly error was added in the previous step:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/132885. It
unblocked the Gitaly version upadte, but it was not the best solution.

Since then, the new version of Gitaly was released and now Rails has
access to structured error content of RawBlame errors.

**Solution**

* Explicitly handle structured errors and raise ArgumentError for
invalid input parameters.
* Add missing tests