Move pages_path_prefix to pages.path_prefix
After some conversation about how to [enable pages multiple] jobs, we
decided to, for now, only rename the `pages_path_prefix` to
`pages.path_prefix`.

`pages_path_prefix` was [introduced] as a _root_ job option, but now
that we the vision that we want to concentrate all `pages` build
options under the same `pages:` option for better organization and
discoverability.

This will not have any impacts as the `pages_path_prefix` options is
currently behind a disabled feature flag and undocumented. The
[documentation will be added on another MR].

[enable pages multiple]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/133624
[introduced]: https://gitlab.com/gitlab-org/gitlab/-/issues/416492
[documentation will be added on another MR]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/130314

Related to:
- https://gitlab.com/gitlab-org/gitlab/-/issues/427755
- https://gitlab.com/gitlab-org/gitlab/-/issues/428018