Fix current_patch_id_sha spec to work with factory default
In `merge_request_spec.rb`, we keep the factory default for project
for all tests. `MergeRequest#current_patch_id_sha`, will call
`project.repository.get_patch_id` internally but we are stubbing
next instance of `Repository`.

For it to work properly with factory default and other tests in
the spec file, we add message expectation on
`merge_request.project.repository` instead as this ensures that
the project factory default was stubbed.