Update the Compliance Center tests
- Rename test methods from 'report' to 'center'.
- Update QA selectors to test-ids.
- Rename UI test-ids on tabs to '*-tab-content' because the ids are on the
element that represents the content, not the tab that users can click.
This leaves '*-tab' test-ids for the clickable tab control, which can be
used as QA selectors.
- Adds a method to switch to the violations tab if it's not already
active. This allows the test to work when the tab is already selected,
as it is when the `adherence_report_ui` feature flag is disabled, and
when the tab is not selected, as when the `adherence_report_ui` feature
flag is enabled.