Increase interruption retries for GitHub workers
https://gitlab.com/gitlab-org/gitlab/-/issues/416777

When the Sidekiq process is asked to stop it will wait for a period of
time for a running job to finish, before forcefully terminating and
requeuing the job if it does not finish within this time.

Normally Sidekiq will do this over and over for a job indefinitely.

The `reliable-fetcher` gem puts a limit on the number of times a job
will be terminated and requeued to 3 times. After this the job will
not be requeued when terminated.

This offers protection from jobs that can never complete in time between
Sidekiq restarts (for example, the time between deploys).

Many of our GitHub importer "stage" workers:

- can work for a very long time as they fetch all the data from GitHub
  to be imported
- pick up from where they left off when interrupted

For these "stage" workers we're increasing the number of times they will
be requeued after being terminated.

https://gitlab.com/gitlab-org/gitlab/-/issues/416777