Change Issue activity to match MRs
This changes all actions in an issue exept for closing to use the small
dot style used on MRs. It also changes the size of the icons in an
issue's activity feed to match those in MRs.

This MR also removes some custom styling that was applied in
merge_requests.scss and upstreams it into issuables.scss.

Changelog: fixed