Update some specs to pass with the super sidebar
This updates several specs to pass when the super sidebar is enabled.

In some cases, we were checking whether the correct nav item was marked
as active in controller/request specs. These test cases have all been
moved to a single feature spec as the Js driver needs to be enabled.

The `have_active_navigation` and `have_active_sub_navigation` custom
matchers have been updated to look for the right selectors in the super
sidebar. Using those matchers now requires the Js driver to be enabled
and should therefore be used in feature specs only.