Update backend for namespaces for emails_enabled
Add a delegate to use the namespace_settings for the emails_enabled
data.

Add functions to namespace_setting to answer the `emails_enabled?`
question by checking all of the ancestor namespaces.

Add parameters to the API which show the new attribute

Update the documentation for the API as well.

Changelog: added