Switch to UrlValidator for Identifer URL validation
AddressableUrlValidator makes a getaddrinfo syscall when validating the
URL. This is a problem for the Vulnerabilities::Identifier model because
we create them in bulk, and the syscall makes a DNS query over network
for non-local URLs. This means that bulk inserts might need to wait for
100 DNS queries (potentially over network) to finish!

To fix this, switch to using UrlValidator instead, which validates URLs
using regex only.

Changelog: fixed
EE: true