Hide repository size limit field on project settings without admin mode
If the admin is an owner of the project or group and admin mode is
enabled for the instance, but the admin has not entered admin mode,
then do not show the repository size limit field on the project or
group settings page. This field requires admin privileges.

Changelog: fixed
EE: true