Refine the etag verification service
Some registries can ommit to return an ETag field on HEAD requests. This
will mess with the dependency proxy cache.
We change this by simply not reseting the cache if the ETag is not
present.
Basically, if the cached file exists and the remote registry doesn't
present an ETag, that situation is considered the same way as for
network errors: we return the cached file.