Make snippet creation resilient to replication lag
When a snippet is being created, it inserts a record in the
database and then performs a Gitaly call to create Git data.

The Gitaly call performs HTTP request for permission check and
if replication lag takes place, those HTTP calls may try fetching
data from a replica that is not up-to-date yet.

This commit:

- Sticks to primary before performing the Gitaly call
- Make the internal API requests aware of the sticking

Changelog: fixed